mistral-eval-platform/requirements.txt
# Core Framework
fastapi>=0.109.0,<1.0.0
uvicorn[standard]==0.27.0
pydantic>=2.8.2,<2.9.0

# Mistral AI SDK
mistralai==1.0.0

# Environment Management
python-dotenv==1.0.0

# HTTP Client (required by mistralai>=1.0.0)
httpx>=0.27.0,<0.28.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==24.1.0
isort==5.13.2
mypy==1.8.0
